db: move Pool from web to db
