+++ src/web/releases.rs
@@ -2,8 +2,8 @@
 
 use super::error::Nope;
 use super::page::Page;
-use super::pool::Pool;
 use super::{duration_to_str, match_version, redirect_base};
+use crate::db::Pool;
 use chrono::{DateTime, NaiveDateTime, Utc};
 use iron::prelude::*;
 use iron::status;
