+++ src/web/rustdoc.rs
@@ -5,9 +5,9 @@ use super::error::Nope;
 use super::file::File;
 use super::metrics;
 use super::page::Page;
-use super::pool::Pool;
 use super::redirect_base;
 use super::{match_version, MatchSemver};
+use crate::db::Pool;
 use crate::utils;
 use crate::Config;
 use iron::headers::{CacheControl, CacheDirective, Expires, HttpDate};
